+++ CHANGELOG.md
@@ -5,11 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## v0.2.0 (2026-07-25)
 
-### Added (for new features.)
-### Changed (for changes in existing functionality.)
-### Deprecated (for soon-to-be removed features.)
-### Removed (for now removed features.)
-### Fixed (for any bug fixes.)
-### Security (in case of vulnerabilities.)
+### Feat
+
+- **pre-commit**: add pre-commit
+
+### Fix
+
+- **discord**: fix error on sending discord message
+
+### Refactor
+
+- **contants**: make colors messenger agnostic
+
+## v0.1.0 (2026-04-27)
+
+### Feat
+
+- **cli**: add the cli
+- **managers**: add dnf manager
+- **managers**: add apt manager
+- **managers**: add base class for managers
+- **messenger**: add discord messenger
+- **logger**: add logger module
+- **models**: add models module